feat: splitting core api functionality out into a separate package
