--- packages/api/src/lib/index.ts
@@ -1,6 +1,0 @@
-import parseResponse from './parseResponse';
-import prepareAuth from './prepareAuth';
-import prepareParams from './prepareParams';
-import prepareServer from './prepareServer';
-
-export { parseResponse, prepareAuth, prepareParams, prepareServer };
