+++ packages/api/src/packageInfo.ts
