+++ packages/core/.eslintignore
@@ -1,0 +1,3 @@
+@types/
+coverage/
+dist/
