+++ packages/core/.eslintrc
@@ -1,0 +1,31 @@
+{
+  "extends": [
+    "@readme/eslint-config",
+    "@readme/eslint-config/docs",
+    "@readme/eslint-config/typescript"
+  ],
+  "root": true,
+  "env": {
+    "es2020": true
+  },
+  "globals": {
+    "fetch": true,
+    "Response": true
+  },
+  "parserOptions": {
+    "ecmaVersion": 2020
+  },
+  "rules": {
+    // @todo fix these eventually
+    "@typescript-eslint/no-explicit-any": "off",
+
+    // Though we aren't requiring JSDoc blocks to be present, if they are they should be properly
+    // formatted.
+    "jsdoc/require-jsdoc": "off",
+
+    // TypeScript types are better suited for these cases instead.
+    "jsdoc/require-param-type": "off",
+    "jsdoc/require-returns": "off",
+    "jsdoc/require-returns-type": "off"
+  }
+}
