+++ packages/core/.gitignore
@@ -1,0 +1,4 @@
+@types/
+coverage/
+dist/
+node_modules/
