+++ packages/core/__tests__/__fixtures__/hello.jp.txt
@@ -1,0 +1,1 @@
+速い茶色のキツネは怠惰な犬を飛び越えます
