+++ packages/core/__tests__/__fixtures__/hello.txt
@@ -1,0 +1,1 @@
+Hello world!
