+++ packages/core/__tests__/parseResponse.test.ts
@@ -1,5 +1,5 @@
 import 'isomorphic-fetch';
-import parseResponse from '../../src/lib/parseResponse';
+import parseResponse from '../src/lib/parseResponse';
 
 const responseBody = JSON.stringify({
   id: 9205436248879918000,
