+++ packages/core/__tests__/prepareAuth.test.ts
@@ -1,5 +1,5 @@
 import Oas from 'oas';
-import prepareAuth from '../../src/lib/prepareAuth';
+import prepareAuth from '../src/lib/prepareAuth';
 
 import securityOas from '@readme/oas-examples/3.0/json/security.json';
