+++ packages/core/package.json
@@ -1,14 +1,13 @@
 {
-  "name": "api",
+  "name": "@readme/api-core",
   "version": "4.2.0",
-  "description": "Generate an SDK from an OpenAPI definition",
+  "description": "Core functionality behind ReadMe's API SDK framework",
   "main": "dist/index.js",
   "types": "@types/index.d.ts",
   "scripts": {
     "build": "tsc",
     "lint": "eslint . --ext .js,.ts",
-    "prebuild": "rm -rf dist/ @types/; npm run prebuild.packageConfig",
-    "prebuild.packageConfig": "node -p \"'export const PACKAGE_NAME = \\'' + require('./package.json').name + '\\';\\nexport const PACKAGE_VERSION = \\'' + require('./package.json').version + '\\';'\" > src/packageInfo.ts",
+    "prebuild": "rm -rf dist/ @types/",
     "prepack": "npm run build",
     "pretest": "npm run lint",
     "prettier": "prettier --list-different --write \"./**/**.{js,ts}\"",
@@ -17,7 +16,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/readmeio/api.git",
-    "directory": "packages/api"
+    "directory": "packages/core"
   },
   "bugs": {
     "url": "https://github.com/readmeio/api/issues"
@@ -28,28 +27,18 @@
     "node": "^14 || ^16"
   },
   "dependencies": {
-    "@readme/api-core": "file:../core",
     "@readme/oas-to-har": "^15.0.0",
-    "@readme/openapi-parser": "^2.0.0",
-    "fetch-har": "^6.1.0",
-    "find-cache-dir": "^3.3.1",
-    "form-data-encoder": "^1.7.1",
-    "formdata-node": "^4.3.2",
+    "datauri": "^4.1.0",
+    "get-stream": "^6.0.0",
     "isomorphic-fetch": "^3.0.0",
-    "js-yaml": "^4.1.0",
-    "make-dir": "^3.1.0",
     "oas": "^17.6.0"
   },
   "devDependencies": {
     "@readme/eslint-config": "^8.4.1",
     "@readme/oas-examples": "^4.4.0",
-    "@types/find-cache-dir": "^3.2.1",
     "@types/jest": "^27.4.0",
-    "@types/js-yaml": "^4.0.5",
     "eslint": "^8.8.0",
     "jest": "^27.4.7",
-    "memfs": "^3.4.1",
-    "nock": "^13.2.2",
     "prettier": "^2.5.1",
     "ts-jest": "^27.1.3",
     "typescript": "^4.7.0-dev.20220216"
