+++ packages/core/src/lib/prepareAuth.ts
