+++ packages/core/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "baseUrl": "./src",
+    "declaration": true,
+    "declarationDir": "./@types",
+    "esModuleInterop": true,
+    "lib": ["dom", "es2020"],
+    "module": "node12",
+    "moduleResolution": "node12",
+    "noImplicitAny": true,
+    "outDir": "dist"
+  },
+  "include": ["./src/**/*"]
+}
